requirements.txt
attrs
tensorflow
keras-tuner
loguru
tqdm